refactor: 토큰 추출 클래스 생성 (중복 방지)
